refactor: renamed newhome to home
