+++ src/templates/Home/DaoCardInfo/styles.ts
