+++ src/templates/Home/HeroHome/HeroBackGround/styles.ts
