+++ src/templates/Home/HeroHome/index.tsx
