+++ src/templates/Home/InvestorsSection/index.tsx
