+++ src/templates/Home/InvestorsSection/styles.ts
