+++ src/templates/Home/KacySection/KacyCardData/styles.ts
