+++ src/templates/Home/LatestNews/NewsCard/styles.ts
