+++ src/templates/Home/ManagerCard/index.tsx
