+++ src/templates/Home/ManagerCard/styles.ts
