+++ src/templates/Home/ManagerSection/styles.ts
