+++ src/templates/Home/PoolHomeCard/styles.ts
