+++ src/templates/Home/SectionCard/styles.ts
