+++ src/templates/Home/TokenIcons/index.tsx
