+++ src/templates/Home/TokenIcons/styles.ts
